[IMP] orm,base: use bigint for "id" primary keys and their foreign keys
The "id" primary keys and the foreign keys referencing them were created
as 32-bit integers (int4/SERIAL). Long-lived databases with high write
churn can exhaust the 2^31-1 space, after which any INSERT on the
affected table fails.

This is not theoretical: in May 2025, quay.io (Red Hat's container
registry) suffered a major outage because the primary key of one of its
tables reached the maximum value of a 32-bit integer, so new rows could
no longer be inserted. See "Quay.io down: primary key has reached
maximum value" https://news.ycombinator.com/item?id=43972181

Changes:
- Many2one and Many2oneReference columns are now int8
- the "id" field column is now int8 and tables are created with
  BIGSERIAL (create_model_table, base_data.sql, the orm signaling tables
  and res.users.apikeys)
- many2many relation tables are created with BIGINT columns
- SQL_ORDER_BY_TYPE orders int8 first (8 bytes aligned on 8 bytes)

Existing databases are migrated in place by the new
base/migrations/1.4/pre-migration.py (base version bumped to 1.4),
executed when updating the base module:
- every int4 "id" column is altered to bigint together with its sequence
  (ALTER SEQUENCE ... AS BIGINT), preserving the nextval default; tables
  using PostgreSQL table inheritance (the ir_actions family) are
  detached, migrated apart and reattached, because ALTER TYPE cannot
  recurse into children whose not-null constraint name differs from the
  parent one
- every int4 column with a foreign key constraint (many2one and
  many2many relation columns) is altered to bigint
- ids remain positive: sequences keep starting at 1, bigint extends the
  headroom up to 2^63-1

Plain Integer fields storing record ids without a FK constraint (e.g.
ir_logging.create_uid, ir.actions.act_window.res_id) are kept as int4
and can be handled separately.

+++ odoo/addons/base/models/res_users.py
@@ -1536,9 +1536,9 @@ class ResUsersApikeys(models.Model):
         table = SQL.identifier(self._table)
         self.env.cr.execute(SQL("""
         CREATE TABLE IF NOT EXISTS %(table)s (
-            id serial primary key,
+            id bigserial primary key,
             name varchar not null,
-            user_id integer not null REFERENCES res_users(id) ON DELETE CASCADE,
+            user_id bigint not null REFERENCES res_users(id) ON DELETE CASCADE,
             scope varchar,
             expiration_date timestamp without time zone,
             index varchar(%(index_size)s) not null CHECK (char_length(index) = %(index_size)s),
