+++ odoo/orm/fields_relational.py
@@ -236,7 +236,7 @@ class Many2one(_Relational):
         company_id(s) are compatible with the currently active company.
     """
     type = 'many2one'
-    _column_type = ('int4', 'int4')
+    _column_type = ('int8', 'int8')
 
     ondelete: OnDelete | None = None    # what to do when value is deleted
     delegate: bool = False              # whether self implements delegation
@@ -1321,8 +1321,8 @@ class Many2many(_RelationalMulti):
         comodel = model.env[self.comodel_name]
         if not sql.table_exists(cr, self.relation):
             cr.execute(SQL(
-                """ CREATE TABLE %(rel)s (%(id1)s INTEGER NOT NULL,
-                                          %(id2)s INTEGER NOT NULL,
+                """ CREATE TABLE %(rel)s (%(id1)s BIGINT NOT NULL,
+                                          %(id2)s BIGINT NOT NULL,
                                           PRIMARY KEY(%(id1)s, %(id2)s));
                     COMMENT ON TABLE %(rel)s IS %(comment)s;
                     CREATE INDEX ON %(rel)s (%(id2)s, %(id1)s); """,
